requirements.txt
asgiref==3.5.2
astroid==2.11.5
autopep8==1.6.0
black==22.3.0
click==8.1.3
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
Django==4.0.5
django-crispy-forms==1.14.0
django-filter==22.1
django-import-export==2.8.0
et-xmlfile==1.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
mccabe==0.7.0
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.10
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pylint==2.14.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
tzdata==2022.1
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
